Many bugfixes and reporting verbosity

+++ robob/driver.py
@@ -1,4 +1,5 @@
 
+import errno
 import time
 import signal
 import select
@@ -20,6 +21,7 @@ class PtyProcess:
 		self.cmdline = cmdline
 		self.kwargs = kwargs
 		self.returncode = None
+		self.fd = None
 
 		# Fork into a new pty
 		self.pid, self.fd = pty.fork()
@@ -57,6 +59,18 @@ class PtyProcess:
 		self.stdin = os.fdopen( fd_write, 'wb', 0 )
 		self.stdout = os.fdopen( fd_read, 'rb', 0 )
 
+	def close(self):
+		"""
+		Close FDs
+		"""
+
+		if self.fd:
+			if self.returncode != None:
+				os.close(self.fd)
+				self.stdin.close()
+				self.stdout.close()
+			self.fd = None
+
 	def send_signal(self, sig):
 		"""
 		Send signal to process
@@ -75,12 +89,13 @@ class PtyProcess:
 		"""
 
 		# Get PID
-		try:
-			pid, sts = os.waitpid(self.pid, os.WNOHANG)
-			if pid == self.pid:
-				self._handle_exitstatus(sts)
-		except os.error as e:
-			self.returncode = None
+		if self.returncode is None:
+			try:
+				pid, sts = os.waitpid(self.pid, os.WNOHANG)
+				if pid == self.pid:
+					self._handle_exitstatus(sts)
+			except os.error as e:
+				self.returncode = None
 
 		return self.returncode
 
@@ -131,31 +146,77 @@ class TestStreamThread(Thread):
 		self.proc = None
 		self.stream = stream
 		self.returncode = 0
+		self.interrupted = False
+		self.interruptReason = ""
+		self.logger = logging.getLogger("stream.%s" % self.stream.name)
 
-	def interrupt(self):
+	def interrupt(self, reason="User interrupt", timeout=5):
 		"""
 		Interrupt the subprocess
 		"""
 
+		# Skip if already interrupted
+		if self.interrupted:
+			return
+
+		# Mark as interrupted
+		self.logger.critical("Interrupting stream")
+		self.interrupted = True
+		self.interruptReason = reason
+
 		# First gracefully kill the sub-process
 		if self.proc:
-			self.proc.send_signal(signal.SIGINT)
+
+			# First send sigint
+			self.proc.send_signal( signal.SIGINT )
+
+			# Wait 5 seconds to 
+			t_timeout = time.time() + timeout
+			while True:
+
+				# Check if this was enough
+				if self.proc.poll() != None:
+					self.logger.warn("Stream interrupted")
+					break
+
+				# Check for timeouts
+				if time.time() > t_timeout:
+
+					# Terminate
+					self.proc.terminate()
+					self.logger.warn("Stream terminated")
+					break
+
+			# Reap process
+			self.proc.close()
+			self.proc = None
+
+			# Set return code to -1
+			self.returncode = -1
 
 	def run(self):
 		"""
 		Main run function
 		"""
 		pipe = self.stream.pipe
-		logger = logging.getLogger("stream.%s" % self.stream.name)
 
 		# Apply delay
 		if self.stream.delay:
-			logger.info("Delaying for %i seconds" % self.stream.delay)
-			time.sleep( self.stream.delay )
+			self.logger.info("Delaying for %i seconds" % self.stream.delay)
+			q_time = time.time() + self.stream.delay
+			while time.time() < q_time:
+				time.sleep(0.1)
+				if self.interrupted:
+					self.logger.warn("Interrupted while in delay")
+					return
 
+		# Prepare expiration time
+		t_expire = None
+		if self.stream.timeout:
+			t_expire = time.time() + self.stream.timeout
 
 		# Open process
-		logger.debug("Process starting %r" % pipe.pipe_cmdline())
+		self.logger.debug("Process starting %r" % pipe.pipe_cmdline())
 		self.proc = proc = PtyProcess( pipe.pipe_cmdline() )
 
 		# Get a sequence of optional expect entries
@@ -165,10 +226,13 @@ class TestStreamThread(Thread):
 
 		# Send stdin if there are no expect entries
 		if (len(expect_out) == 0) and (len(expect_err) == 0):
-			logger.debug("Sending STDIN payload")
+			self.logger.debug("Sending STDIN payload")
 			os.write( proc.fd, pipe.pipe_stdin() )
 			os.write( proc.fd, "\x04") # End-of-transmission
 			self.has_expect = False
+
+		# Flag that keeps line polling activ
+		active = True
 
 		# Helper function to handle lines
 		def handle_line(read):
@@ -178,14 +242,14 @@ class TestStreamThread(Thread):
 				return
 
 			# Guard against exceptions
-			logger.debug("STDIN: %s" % read)
+			self.logger.debug("STDIN: %s" % read)
 			try:
 
 				# First apply over expect
 				handled = False
 				for i in range(0, len(expect_out)):
 					if expect_out[i].matches( read ):
-						logger.debug("Expect matched as /%s/ on stdin" % str(expect_out[i]))
+						self.logger.debug("Expect matched as /%s/ on stdin" % str(expect_out[i]))
 						os.write( proc.fd, expect_out[i].render() )
 						del expect_out[i]
 						handled = True
@@ -199,86 +263,73 @@ class TestStreamThread(Thread):
 
 				# In case something went wrong while processing the streams,
 				# kill the process
-				logger.error("%s occured: %s" % (e.__class__.__name__, str(e)) )
+				self.logger.error("%s occured: %s" % (e.__class__.__name__, str(e)) )
+				active = False
 
-				# First send sigint
-				proc.send_signal( signal.SIGINT )
-
-				# Wait 5 seconds to 
-				timeout = time.time() + 5
-				while True:
-
-					# Check if this was enough
-					if proc.poll() != None:
-						logger.warn("Stream interrupted")
-						proc.stdout.close()
-						self.proc = None
-						self.returncode = -1
-						return
-
-					# Check for timeouts
-					if time.time() > timeout:
-
-						# Terminate
-						proc.terminate()
-						logger.warn("Stream terminated")
-
-						# Cleanup
-						proc.stdout.close()
-						self.proc = None
-						self.returncode = -1
-						return
+				# Interrupt
+				self.interrupt("%s: %s" % (e.__class__.__name__, str(e)))
+				return
 
 			# If we have processed all expect entries, send STDIN
 			if self.has_expect and (len(expect_out) == 0) and (len(expect_err) == 0):
-				logger.debug("No more expects left, sending STDIN payload")
+				self.logger.debug("No more expects left, sending STDIN payload")
 				os.write( proc.fd, pipe.pipe_stdin() )
 				os.write( proc.fd, "\x04") # End-of-transmission
 				self.has_expect = False
 
 		# Read stdout/err
-		logger.debug("Processing output")
+		self.logger.debug("Processing output")
 		data = ""
-		while True:
+		data_flush_t = 0
+		while active:
 
-			# Read a chunk of data
-			buf = os.read( proc.fd, 1024)
-			if buf:
+			# If interrupted, quit
+			if self.interrupted:
+				return
 
-				# Process lines
-				data += buf
-				while "\n" in data:
-					(line, data) = data.split("\n",1)
-					handle_line(line)
-
-				# If there are more data don't consider incomplete
-				# lines as part of the output
-				if len(buf) < 1024:
-					handle_line(data)
-					data = ""
-
-			elif data:
-
-				# Consider this an incomplete line that just finished
+			# Forward incomplete data as incomplete line
+			if data and (time.time() > data_flush_t):
 				handle_line(data)
 				data = ""
 
+			# Wait for data within 100ms
+			if proc.fd in select.select([proc.fd], [], [], 0.1)[0]:
+				buf = os.read( proc.fd, 4096)
+				if buf:
+
+					# Process lines
+					data += buf
+					while "\n" in data:
+						(line, data) = data.split("\n",1)
+						handle_line(line)
+
+					# Note when we received the data
+					# in order to forward them as-is as
+					# incomplete line data after a timeout
+					data_flush_t = time.time() + 0.1
+
 			# Check for process exit
 			if proc.poll() != None:
-				logger.debug("Process exited with code %i" % proc.returncode)
+				self.logger.debug("Process exited with code %i" % proc.returncode)
 				break
 
+			# Check if timeout expired
+			if t_expire and (time.time() > t_expire):
+				self.logger.critical("Timeout of %s seconds expired" % self.stream.timeout)
+				self.interrupt("Timeout after %s sec" % self.stream.timeout)
+				return
 
-		# Clear
-		proc.stdout.close()
+		# If not interrupted, clean shutdown
+		if not self.interrupted and self.proc:
+			proc.close()
+			self.returncode = proc.returncode
+			self.proc = None
 
 		# Cleanup
-		self.proc = None
-		self.returncode = proc.returncode
-		if self.returncode > 0:
-			logger.warn("Stream exited with code=%i" % proc.returncode)
+		if self.returncode != 0:
+			self.logger.warn("Stream closed with error code=%i" % proc.returncode)
 		else:
-			logger.debug("Stream exited")
+			self.logger.info("Stream closed successfuly")
 
 class TestDriver:
 	"""
@@ -294,44 +345,79 @@ class TestDriver:
 
 		self.specs = specs
 		self.test = test
+		self.logger = logging.getLogger("driver")
 
 		self.metrics = Metrics()
 		self.metrics.configure( test )
 
 		self.results = []
 		self.lastResults = None
+		self.lastStatus = ""
 
 	def run(self):
 		"""
 		Start the tests on the test driver
 		"""
-		logger = logging.getLogger("driver")
 
 		# Create all stream specifications on this context
 		streams = self.specs.createStreams( self.test, self.metrics )
 
 		# Reset metrics
 		self.metrics.reset()
+		self.lastStatus = "Completed"
 
 		# Launch them
-		threads = []
-		logger.debug("Starting %i streams" % len(streams))
+		self.threads = []
+		self.logger.debug("Starting %i streams" % len(streams))
 		for s in streams:
+
+			# Skip inactive streams
+			if not s.active:
+				self.logger.info("Not starting inactive stream '%s'" % s.name)
+				continue
 
 			# Create a test stream thread
 			t = TestStreamThread( s )
 
 			# Start and keep
 			t.start()
-			threads.append(t)
+			self.threads.append(t)
 
-		# Wait for the to complete
-		logger.debug("Waiting for stream threads to exit")
-		for t in threads:
+		# Wait for all threads to complete
+		self.logger.debug("Waiting for stream threads to exit")
+		hasAlive = True
+		hasInterrupt = False
+		while hasAlive:
+
+			# Check status of all items
+			hasAlive = False
+			for t in self.threads:
+				if t.interrupted:
+					hasInterrupt = True
+					break
+				if t.is_alive():
+					hasAlive = True
+					break
+
+			# If something interrupted, interrupt everything
+			if hasInterrupt:
+				self.logger.warn("Thread '%s' interrupted, so collapsing the test" % t.stream.name)
+				self.interrupt(t.interruptReason)
+				break
+
+			# Otherwise wait for a sec before retry
+			if hasAlive:
+				time.sleep(0.1)
+
+		# Reap threads
+		for t in self.threads:
 			t.join()
 
+		# Reset thread list
+		self.threads = []
+
 		# Collect results
-		logger.debug("Threads exited, collecting results")
+		self.logger.debug("Threads exited, collecting results")
 		self.lastResults = self.metrics.results()
 		self.results.append( self.lastResults )
 
@@ -343,4 +429,20 @@ class TestDriver:
 		# Summarize results
 		return summarize( self.results )
 
+	def interrupt(self, reason="Interrupted"):
+		"""
+		Interrupt current run
+		"""
 
+		# Collect results
+		self.logger.warn("Driver interrupted (%s), stopping all threads" % reason)
+		self.lastStatus = reason
+		self.lastResults = self.metrics.results()
+		self.results.append( self.lastResults )
+
+		# Interrupt and join all threads
+		for t in self.threads:
+			self.logger.debug("Interrupting stream thread %s" % t.stream.name)
+			t.interrupt()
+			self.logger.debug("Joining stream thread %s" % t.stream.name)
+			t.join()
